Initial project restructure
