+++ lib/constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+    OVERRIDE_DIR: 'overrides',
+    GENERATED_DIR: 'generated',
+    LISTS_DIR: 'features'
+}
