+++ lib/platforms.js
